Move GeniusApiService to Lib.Services
