+++ src/Lib.Services/Services/GeniusApiService/GeniusApiServiceExtensions.cs
@@ -1,7 +1,7 @@
 using System.Reflection;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace MuzakBot.App.Services;
+namespace MuzakBot.Lib.Services;
 
 /// <summary>
 /// Extension methods for adding the Genius API service to the service collection.
